requirements.txt
yt-dlp>=2023.0.0
prometheus-client>=0.19.0
psutil>=5.9.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
flask>=2.3.0
google-api-python-client>=2.0.0
google-auth>=2.0.0